requirements.txt
pyannote.metrics
simpleder
